requirements.txt
django>=5.1
django-allauth
django-cors-headers
django-health-check 
djangorestframework
drf-spectacular
pillow
psycopg
